moved into same file as Predictive, added arbitrary batch shape testing
