+++ pyro/infer/__init__.py
@@ -11,7 +11,7 @@ from pyro.infer.importance import Importance
 from pyro.infer.mcmc.api import MCMC
 from pyro.infer.mcmc.hmc import HMC
 from pyro.infer.mcmc.nuts import NUTS
-from pyro.infer.predictive import Predictive
+from pyro.infer.predictive import Predictive, log_likelihood
 from pyro.infer.renyi_elbo import RenyiELBO
 from pyro.infer.rws import ReweightedWakeSleep
 from pyro.infer.smcfilter import SMCFilter
@@ -42,6 +42,7 @@ __all__ = [
     "JitTraceGraph_ELBO",
     "JitTraceMeanField_ELBO",
     "JitTrace_ELBO",
+    "log_likelihood",
     "MCMC",
     "NUTS",
     "Predictive",
